Updated exp_export_line_data.py to generate a running total of points earned.
Copied the experiment to the main directory as export_line_data.py, and
added the HTML files and team_colors.js to render the data.

finishes.html will display the finishing position of each team each
season

points_per_season.html will display the points earned by each team for
each season

points_running_total.html will display the running total through the
years

NB: team_colors.js contains color values from
https://encycolorpedia.com/teams/football
so that each team's line is represented by their team's primary dark
color.  It is a bit of a hack, as this is a static file, it really
should be procedurally generated from a table in the database so it
doesn't fall out of date as new teams enter the league.